_posts/Cuda/actual_001_reverse_curved_scanconversion/requirements.txt
matplotlib==3.7.3
natsort==8.4.0
numpy==1.24.3
Pillow==10.2.0
pydicom==2.4.4
pynrrd==1.0.0
# PyQt5==5.15.10
PyQt5
# scipy==1.11.4
scipy
